requirements.txt
# Core dependencies
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Vector database
faiss-cpu==1.7.4
numpy==1.24.3

# Text processing
mecab-python3==1.0.6
unidic-lite==1.0.8
spacy==3.7.2
fugashi==1.3.0

# Database
sqlalchemy==2.0.23
sqlite-utils==3.35.2

# API framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# UI
gradio==4.8.0
streamlit==1.29.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1